init merge attempt
