--- Group Project/game_platform/Pipfile
@@ -1,14 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-flask = "*"
-pymysql = "*"
-flask-bcrypt = "*"
-
-[dev-packages]
-
-[requires]
-python_version = "3.8"
