--- Group Project/game_platform/flask_app/__init__.py
@@ -1,5 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-app.secret_key = "SHHHHHHHH"
