--- Group Project/game_platform/flask_app/models/user.py
@@ -1,70 +1,0 @@
-from flask_app.config.mysqlconnection import connectToMySQL
-import re
-EMAIL_REGEX = re.compile(r'^[a-zA-Z0-9.+_-]+@[a-zA-Z0-9._-]+\.[a-zA-Z]+$')
-from flask import flash
-
-class User:
-    db_name = "game_platform"
-    def __init__(self,data):
-        self.id = data['id']
-        self.username = data['username']
-        self.email = data['email']
-        self.password = data['password']
-        self.created_at = data['created_at']
-        self.updated_at = data['updated_at']
-
-    @classmethod
-    def save(cls,data):
-        query = "INSERT INTO users (username,email,password) VALUES(%(username)s,%(email)s,%(password)s)"
-        return connectToMySQL(cls.db_name).query_db(query,data)
-
-    @classmethod
-    def get_all(cls):
-        query = "SELECT * FROM users;"
-        results = connectToMySQL(cls.db_name).query_db(query)
-        users = []
-        for row in results:
-            users.append( cls(row))
-        return users
-
-    @classmethod
-    def get_by_email(cls,data):
-        query = "SELECT * FROM users WHERE email = %(email)s;"
-        results = connectToMySQL(cls.db_name).query_db(query,data)
-        if len(results) < 1:
-            return False
-        return cls(results[0])
-
-    @classmethod
-    def get_by_id(cls,data):
-        query = "SELECT * FROM users WHERE id = %(id)s;"
-        results = connectToMySQL(cls.db_name).query_db(query,data)
-        return cls(results[0])
-    
-    @classmethod
-    def update(cls, data):
-        query = "UPDATE users SET username=%(username)s WHERE id = %(id)s;"
-        return connectToMySQL(cls.db_name).query_db(query,data)
-    
-
-    @staticmethod
-    def validate_register(user):
-        is_valid = True
-        query = "SELECT * FROM users WHERE email = %(email)s;"
-        results = connectToMySQL(User.db_name).query_db(query,user)
-        if len(results) >= 1:
-            flash("Email already taken.","register")
-            is_valid=False
-        if not EMAIL_REGEX.match(user['email']):
-            flash("Invalid Email","register")
-            is_valid=False
-        if len(user['username']) < 3:
-            flash("First name must be at least 3 characters","register")
-            is_valid= False
-
-        if len(user['password']) < 8:
-            flash("Password must be at least 8 characters","register")
-            is_valid= False
-        if user['password'] != user['confirm']:
-            flash("Passwords don't match","register")
-        return is_valid
