--- Group Project/game_platform/flask_app/templates/dashboard.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Dashboard</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-</head>
-
-<body style="background-color:lightgray">
-    <nav class="d-flex justify-content-between align-items-center col-12">
-        <h1 class="text-center">Game Platform</h1>
-        <h2>Welcome {{user.username}}</h2>
-        <div class="d-flex col-4 justify-content-around">
-            <a href="/edit/user/{{user.id}}" class="btn btn-info">edit profile</a>
-            <a href="/logout" class="btn btn-info">Logout</a>
-
-        </div>
-
-    </nav>
-    <div class="col-2 mx-auto" style="background-color:darkkhaki">
-        <h2>Users:</h2>
-        <table class="table table-hover">
-            <tbody>
-                {% for user in users %}
-                <tr>
-                    <p>{{user.username}}</p>
-                </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-</body>
-
-</html>
