--- Group Project/game_platform/flask_app/templates/edit_user.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Edit Profile</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-</head>
-
-<body style="background-color:lightgray">
-    <nav class="d-flex justify-content-between align-items-center col-12">
-        <h1 class="text-center">Game Platform</h1>
-        <div class="d-flex col-6 justify-content-around">
-            <a href="/dashboard" class="btn btn-primary">Dashboard</a>
-            <a href="/logout" class="btn btn-info">Logout</a>
-        </div>
-
-    </nav>
-    <form action="/update/user" method="post" class="col-11 p-4 my-4 mx-auto">
-        {% with messages = get_flashed_messages(category_filter=['user']) %}
-        {% if messages %}
-        <div class="alert" role="alert">
-            <strong>alert</strong>
-            {% for message in messages %}
-            <p>{{message}}</p>
-            {% endfor %}
-        </div>
-        {% endif %}
-        {% endwith %}
-
-        <div class="d-flex justify-content-between col-50">
-            <input type="hidden" name="id" value={{edit.id}}>
-            <div class="col-50">
-                <div class="form-group">
-                    <label for="username">Username:</label>
-                    <textarea name="username" cols="20" rows="1" class="form-control">{{edit.username}}</textarea>
-                </div>
-                <div class="form-group">
-                    <label for="email">Email:</label>
-                    <textarea name="email" cols="50" rows="1" class="form-control">{{edit.email}}</textarea>
-                </div>
-                <br>
-                <input type="submit" value="Update" class="btn btn-primary">
-            </div>
-
-        </div>
-
-    </form>
-</body>
-
-</html>
