--- Group Project/game_platform/flask_app/templates/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-</head>
-
-<body style="background-color:lightgray">
-    <div>
-        <form action="/login" method="post" class="col-4 p-4 mx-auto">
-            <h2 class="text-success">Login</h2>
-            {% with messages = get_flashed_messages(category_filter=['login']) %}
-            {% if messages %}
-            <div class="alert alert-danger" role="alert">
-                <strong>alert</strong>
-                {% for message in messages %}
-                <p>{{message}}</p>
-                {% endfor %}
-            </div>
-            {% endif %}
-            {% endwith %}
-            <div class="form-group">
-                <label for="email">Email:</label>
-                <input type="text" name="email" class="form-control">
-            </div>
-            <div class="form-group">
-                <label for="password">Password:</label>
-                <input type="password" name="password" class="form-control">
-            </div>
-
-            <input type="submit" value="Login" class="btn btn-success">
-            <div>
-                <a href="/new">Click here to register</a>
-            </div>
-        </form>
-
-        
-    </div>
-</body>
-
-</html>
