--- Group Project/game_platform/server.py
@@ -1,6 +1,0 @@
-from flask_app import app
-
-from flask_app.controllers import users
-
-if __name__=="__main__":
-    app.run(debug=True)
