+++ flask_app/templates/dashboard.html
@@ -1,17 +1,29 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <title>Game Platform</title>
 </head>
-<body>
-    <!-- Navigation bar -->
+
+<body style="background-color:lightgray">
+    <nav class="d-flex justify-content-between align-items-center col-12">
+        <h1 class="text-center">Game Platform</h1>
+        <h2>Welcome {{user.username}}</h2>
+        <div class="d-flex col-4 justify-content-around">
+            <a href="/edit/user/{{user.id}}" class="btn btn-info">edit profile</a>
+            <a href="/logout" class="btn btn-info">Logout</a>
+        </div>
+    </nav>
     <!-- 'Featured Games' section [Optional] -->
     <!-- 'All Games' section [PL] -->
     <!-- 'All Users' section [HS] -->
-    <div>
+    <div class="col-2 mx-auto" style="background-color:darkkhaki">
+        <h2>Users:</h2>
         {% for user in all_users %}
             <!-- User card -->
             <div>
@@ -22,9 +34,9 @@
             <!-- Make some sort of conditional so that you cannot add yourself to your friends list -->
         {% endfor %}
     </div>
-
     <!-- 'Collection' section [HS] -->
     <!-- 'Wishlist' section [HS] -->
     <!-- 'Friends' section [HS] -->
 </body>
+
 </html>
